Updates to 4.0.7 for graphene.  Still needs work on the property graph generation.  Interaction graph is working in Enron sample.

+++ graphene-parent/graphene-model/src/main/java/graphene/model/query/PageableQuery.java
@@ -5,23 +5,23 @@ public interface PageableQuery {
 	/**
 	 * @return the firstResult
 	 */
-	public abstract int getFirstResult();
+	public abstract long getFirstResult();
 
 	/**
 	 * @return the maxResult
 	 */
-	public abstract int getMaxResult();
+	public abstract long getMaxResult();
 
 	/**
 	 * @param firstResult
 	 *            the firstResult to set
 	 */
-	public abstract void setFirstResult(int firstResult);
+	public abstract void setFirstResult(long firstResult);
 
 	/**
 	 * @param maxResult
 	 *            the maxResult to set
 	 */
-	public abstract void setMaxResult(int maxResult);
+	public abstract void setMaxResult(long maxResult);
 
 }
